Add checksumInFilename option.
When this setting is set to true, the checksum is included as part of
the bundle filename, e.g.
jquery-address-1d3f939fe75ebe1e8975f1728c0925ee.js . The first line of
the versions file dropped into the /resources/bundles directory now
includes the setting of checksum-in-filename as:

checksum-in-filename=true|false

Based on that, the bundle URI can be constructed either with the version
in the filename or after the filename in a querystring parameter.

Originally requested in issue #2.