Allow nested bundle references.
When defining a bundle, you can now reference another bundle by name and
its files will be expanded in place of the bundle name. Bundle names are
distinguished as having no .s in their names. Notably, you can have a
bundle with .s in its name, but you won't be able to mix it into another
bundle.

This fulfills the wish expressed in issue #5.