Pass baseDirectory to coffee and less compilation commands.
Before we were running them based on the current working directory; however, in
situations where sbt-resource-management is attached to subprojects, the base
directory for the commands should be the subproject directories rather than the
project directory. This particular bug had been corrected for compass, but not for
coffee and less.