Add the ability to inject the starting environment for testing.
Because of the way Maps work (I assume) we were never getting the same
array twice, making our equality checks fail. The arrays had the same
contents, but not the same order. Because we can't use argument matchers
here (due to the overloaded-ness of exec), we have to hack around it.